changed method displayRunningTime to take parameter as double instead of int

+++ PIC_Simulator_GUI/src/pic_simulator_gui/presenter/MainWindowPresenterImpl.java
@@ -104,7 +104,7 @@ public class MainWindowPresenterImpl
     }
 
     @Override
-    public void displayRunningTime(int microSeconds) {
+    public void displayRunningTime(double microSeconds) {
         _view.displayRunningTime(microSeconds);
     }
